chore(schema-validation): update react components to functional from classes
