+++ packages/compass-schema-validation/package.json
@@ -85,7 +85,6 @@
     "lodash": "^4.17.21",
     "mongodb-ns": "^2.4.2",
     "mongodb-query-parser": "^4.3.0",
-    "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-redux": "^8.1.3",
     "redux": "^4.2.1",
