--- packages/compass-schema-validation/src/components/compass-schema-validation/index.ts
@@ -1,6 +1,0 @@
-import MappedCompassSchemaValidation, {
-  CompassSchemaValidation,
-} from './compass-schema-validation';
-
-export default MappedCompassSchemaValidation;
-export { CompassSchemaValidation };
