--- packages/compass-schema-validation/src/components/sample-documents/index.ts
@@ -1,4 +1,0 @@
-import SampleDocuments from './sample-documents';
-
-export default SampleDocuments;
-export { SampleDocuments };
