--- packages/compass-schema-validation/src/components/validation-editor/index.ts
@@ -1,4 +1,0 @@
-import ValidationEditor from './validation-editor';
-
-export default ValidationEditor;
-export { ValidationEditor };
