--- packages/compass-schema-validation/src/components/validation-states/index.ts
@@ -1,4 +1,0 @@
-import ValidationStates from './validation-states';
-
-export default ValidationStates;
-export { ValidationStates };
