+++ packages/compass-schema-validation/src/modules/index.ts
@@ -77,7 +77,7 @@ export type SchemaValidationThunkAction<
 > = ThunkAction<R, RootState, SchemaValidationExtraArgs, A>;
 
 /**
- * The intial state of the root reducer.
+ * The initial state of the root reducer.
  */
 export const INITIAL_STATE: RootState = {
   namespace: NS_INITIAL_STATE,
@@ -112,9 +112,6 @@ const doReset = (): RootState => ({ ...INITIAL_STATE });
  */
 export const reset = (): ResetAction => ({ type: RESET });
 
-/**
- * The root reducer.
- */
 const rootReducer = (
   state: RootState | undefined,
   action: RootAction
